test(redis-storage-service): review feedback and improve test

+++ src/cores/redis-storage-service.spec.ts
@@ -7,18 +7,13 @@
  *
  * @copyright (C) 2020 LemonCloud Co Ltd. - All Rights Reserved.
  */
-import { expect2, GeneralItem, GETERR, StorageModel } from '..';
+import { expect2, GETERR, StorageModel } from '..';
 import { RedisStorageService, DummyRedisStorageService } from './redis-storage-service';
+import { isLocalCacheAvailable } from './cache-service.spec';
 
-// export const instance = (type?: 'dummy') => {
-//     if (type === 'dummy') {
-//         return { service: new DummyRedisStorageService() };
-//     } else {
-//         return { service: new RedisStorageService({ endpoint: 'localhost:6379', tableName: 'Test' }) };
-//     }
-// };
 interface TestModel extends StorageModel {
     type?: 'test';
+    B?: boolean;
     N?: number;
     S?: string;
     M?: object;
@@ -36,10 +31,37 @@ describe('RedisStorageService', () => {
         done();
     });
 
+    it('serialize/deserialize', async done => {
+        const service = new DummyRedisStorageService<TestModel>();
+        let model: TestModel;
+        let data: Record<string, string>;
+        /* eslint-disable prettier/prettier */
+
+        model = { type: 'test', B: true, N: 100, S: 'lemon', M: { a: 9, b: 'cloud' }, L: [1, 3, 5, 'redis', 'storage', { empty: null }] };
+        data = service.serialize(model);
+        expect2(() => service.deserialize(data)).toEqual(model);
+
+        // null values must be preserved
+        model = { type: 'test', S: null };
+        data = service.serialize(model);
+        expect2(() => service.deserialize(data)).toEqual(model);
+
+        // undefined values must be removed
+        model = { type: 'test', N: undefined };
+        data = service.serialize(model);
+        expect2(() => data).toEqual({ type: '"test"' });
+        expect2(() => service.deserialize(data)).toEqual({ type: 'test' });
+
+        /* eslint-enable prettier/prettier */
+        done();
+    });
+
     it('CRUD', async done => {
+        if (!(await isLocalCacheAvailable('redis'))) return done(); // dummy-redis-storage-service uses local redis server
         const service = new DummyRedisStorageService<TestModel>();
         /* eslint-disable prettier/prettier */
 
+        // Empty dummy storage
         await service.truncate();
 
         // readOrCreate - will be created
@@ -62,7 +84,7 @@ describe('RedisStorageService', () => {
         // increment w/ update
         expect2(await service.increment('b', { N: 10 }, { S: 'lemon' }).catch(GETERR)).toEqual({ id: 'b', type: 'test', N: 20, S: 'lemon', L: ['foo', 'bar'] });
         // failed: increment for non-numeric field
-        expect2(await service.increment('b', { S: 1 } as any, { N: 0 }).catch(GETERR)).toEqual(`@increment tries to increment non-numeric field [S].`);
+        expect2(await service.increment('b', { S: 1 } as any, { N: 0 }).catch(GETERR)).toEqual(`.S is non-numeric field and cannot be incremented.`);
 
         // failed: existing model will not be changed
         expect2(await service.read('b').catch(GETERR), 'N').toEqual({ N: 20 });
